+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 - Added new example `example-modem-ctrl` to demonstrate how to send control commands of cell IDs and IMSI to the client using the control interface.
 - Fixed a bug where `/CID` (or any update assistance data request) would leak memory if the client couldn't encode the message.
 - Updated modem library and removed it from `example-lpp`. Use control commands to update cell IDs and IMSI instead.
+- Added options to use the SLP address instead of providing `--host`.
+    - `--slp-host-imsi` to use the IMSI to generate the SLP address.
+    - `--slp-host-cell` to use the cell information to generate the SLP address.
 
 ## [3.4.9] 2024-05-03
 - Added a few options to controll how SPARTN messages are generated: