+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Added TCP and UDP support when using the NMEA receiver.
     - `--nmea-tcp` to specify the TCP ip address to connect to.
     - `--nmea-tcp-port` to specify the TCP port to connect to.
+- The encoding of ha-uncertainty has been changed to the closed value and will never be encoded as 0.0m.
 
 ## [3.4.9] 2024-05-03
 - Added a few options to controll how SPARTN messages are generated: